[alpha] Move BossComponent in subfolder Boss.
