+++ code/Alpha/Alpha.vcxproj
@@ -102,7 +102,7 @@
     <ClCompile Include="Bullets\BulletSubsystem.cpp" />
     <ClCompile Include="Bullets\IBulletWave.cpp" />
     <ClCompile Include="Bullets\Waves\WaveTest.cpp" />
-    <ClCompile Include="Components\BossComponent.cpp" />
+    <ClCompile Include="Components\Boss\BossComponent.cpp" />
     <ClCompile Include="Components\PlayerComponent.cpp" />
     <ClCompile Include="Configuration.cpp" />
     <ClCompile Include="main.cpp" />
@@ -416,7 +416,7 @@
     <ClInclude Include="Bullets\BulletSubsystem.h" />
     <ClInclude Include="Bullets\IBulletWave.h" />
     <ClInclude Include="Bullets\Waves\WaveTest.h" />
-    <ClInclude Include="Components\BossComponent.h" />
+    <ClInclude Include="Components\Boss\BossComponent.h" />
     <ClInclude Include="Components\PlayerComponent.h" />
     <ClInclude Include="Configuration.h" />
     <ClInclude Include="Reflection\ReflectionGameTypes.h" />
