+++ code/Alpha/Components/Boss/BossComponent.cpp
@@ -2,7 +2,7 @@
 /* Copyright (C) 2026 Florent Devillechabrol <florent.devillechabrol@gmail.com>	*/
 /********************************************************************************/
 
-#include "Alpha/Components/BossComponent.h"
+#include "Alpha/Components/Boss/BossComponent.h"
 
 #include "Alpha/Bullets/BulletSubsystem.h"
 #include "Alpha/Bullets/Waves/WaveTest.h"
